add all tests for homework3
