--- main.py
@@ -1,63 +1,0 @@
-"""
-Програма для расчета по ДЗ №3
-Версия 1.0
-Автор: Андрей Корниенко
-
-Домашнее задание
-Тестирование API
-
-Цель:
-Поучиться тестировать API сервисов на основе их документации.
-
-Тестирование каждого api оформить в отдельном тестовом модуле.
-
-Тестирование REST сервиса 1
-Написать минимум 5 тестов для REST API сервиса: https://dog.ceo/dog-api/.
-Как минимум 2 из 5 должны использовать параметризацию.
-Документация к API есть на сайте.
-Тесты должны успешно проходить.
-
-Тестирование REST сервиса 2
-Написать минимум 5 тестов для REST API сервиса: https://www.openbrewerydb.org/.
-Как минимум 2 из 5 должны использовать параметризацию.
-Документация к API есть на сайте.
-Тесты должны успешно проходить.
-
-Тестирование REST сервиса 3.
-Написать минимум 5 тестов для REST API сервиса: https://jsonplaceholder.typicode.com/.
-Как минимум 2 из 5 должны использовать параметризацию.
-Документация к API есть на сайте.
-Тесты должны успешно проходить.
-
-Реализуйте в отдельном модуле (файле) тестовую функцию которая будет принимать 2 параметра:
-url - должно быть значение по умолчанию https://ya.ru
-status_code - значение по умолчанию 200
-Параметры должны быть реализованы через pytest.addoption.
-Можно положить фикcтуру и тестовую функцию в один файл.
-Основная задача чтобы ваш тест проверял по переданному урлу статус ответа тот который передали,
-т.е. по адресу https://ya.ru/sfhfhfhfhfhfhfhfh должен быть валидным ответ 404
-
-пример запуска pytest test_module.py --url=https://mail.ru --status_code=200
-
-Критерии оценки:
-Все перечисленные пункты сдавать одним pull-request'ом
-Для всех файлов соблюдается минимальный код сатйл (встроенный форматтер PyCharm'а)
-Под тесты каждого сервиса заведён отдельный файл
-Рекомендуем сдать до: 28.07.2021
-"""
-
-if __name__ == "__main__":
-    import requests
-    import pytest
-
-    proxy = {"http": "localhost:8080", "https": "localhost:8080"}
-    response = requests.get('https://ya.ru', proxies=proxy, verify=False)
-    print(f"response = {response}")
-    assert response.status_code == 200
-
-    print(f"response.headers.items() = {response.headers.items()}")
-    print(f"response.headers.items() = {response.headers.items()}")
-
-    print(100 * '*')
-    for key, value in response.headers.items():
-        print(key, '=>', value)
