+++ tests/conftest.py
@@ -2,31 +2,6 @@ import pytest
 import requests
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        "--url",
-        default="https://ya.ru/",
-        help="This is request url"
-    )
-
-    parser.addoption(
-        "--method",
-        default="get",
-        choices=["get", "post", "put", "patch", "delete"],
-        help="method to execute"
-    )
-
-
-@pytest.fixture
-def base_url(request):
-    return request.config.getoption("--url")
-
-
-@pytest.fixture
-def request_method(request):
-    return getattr(requests, request.config.getoption("--method"))
-
-
 @pytest.fixture
 def http_method_get():
     return getattr(requests, "get")
