+++ tests/test_dog_ceo/test_dog_ceo.py
@@ -1,7 +1,31 @@
+"""
+Програма для расчета по ДЗ №3
+Версия 1.0
+Автор: Андрей Корниенко
+
+Домашнее задание
+Тестирование API
+
+Цель:
+Поучиться тестировать API сервисов на основе их документации.
+
+Тестирование каждого api оформить в отдельном тестовом модуле.
+
+Тестирование REST сервиса 1
+Написать минимум 5 тестов для REST API сервиса: https://dog.ceo/dog-api/.
+Как минимум 2 из 5 должны использовать параметризацию.
+Документация к API есть на сайте.
+Тесты должны успешно проходить.
+
+Критерии оценки:
+Все перечисленные пункты сдавать одним pull-request'ом
+Для всех файлов соблюдается минимальный код сатйл (встроенный форматтер PyCharm'а)
+Под тесты каждого сервиса заведён отдельный файл
+Рекомендуем сдать до: 28.07.2021"""
+
 import json
 import collections
 import pytest
-
 from tests.test_dog_ceo.test_data.test_data import breeds
 
 TEST_DIR_PATH = './test_dog_ceo/download_images'
